+++ string.c
@@ -121,16 +121,16 @@ DEFUN(stringp, "string?", 1, 0, NAVI_ANY)
 DEFUN(make_string, "make-string", 1, NAVI_PROC_VARIADIC, NAVI_FIXNUM)
 {
 	UChar32 ch = ' ';
-	int32_t length = navi_fixnum(scm_arg1);
+	intptr_t length = navi_fixnum(scm_arg1);
 	if (length < 0)
 		navi_error(scm_env, "invalid length");
 	if (scm_nr_args > 1)
 		ch = navi_char_cast(scm_arg2, scm_env);
 
-	int32_t size = length * u8_length(ch);
+	intptr_t size = length * u8_length(ch);
 	navi_obj obj = navi_make_string(size, size, length);
 	struct navi_string *str = navi_string(obj);
-	for (int32_t i = 0; i < size;) {
+	for (intptr_t i = 0; i < size;) {
 		u8_append(str->data, i, size, ch);
 	}
 	return obj;
@@ -195,7 +195,7 @@ void navi_string_set(struct navi_string *str, long k, UChar32 new_ch)
 
 DEFUN(string_set, "string-set!", 3, 0, NAVI_STRING, NAVI_FIXNUM, NAVI_CHAR)
 {
-	int32_t k;
+	intptr_t k;
 	struct navi_string *str;
 
 	str = navi_string(scm_arg1);
@@ -420,7 +420,7 @@ DEFUN(string_copy_to, "string-copy!", 3, NAVI_PROC_VARIADIC,
 		NAVI_STRING, NAVI_FIXNUM, NAVI_STRING)
 {
 	navi_obj to, from;
-	long at, start, end;
+	intptr_t at, start, end;
 
 	to = scm_arg1;
 	at = navi_fixnum(scm_arg2);
